Move replacement operators implementations behind feature gate
